@@ -1,13 +1,14 @@
 name: image_picker
 description: Flutter plugin for selecting images from the Android and iOS image
   library, and taking new pictures with the camera.
-repository: https://github.com/flutter/plugins/tree/main/packages/image_picker/image_picker
+repository: https://github.com/flutter/packages/tree/main/packages/image_picker/image_picker
 issue_tracker: https://github.com/flutter/flutter/issues?q=is%3Aissue+is%3Aopen+label%3A%22p%3A+image_picker%22
-version: 0.8.6
+version: 1.0.4
+publish_to: none
 
 environment:
-  sdk: ">=2.14.0 <3.0.0"
-  flutter: ">=2.10.0"
+  sdk: ">=2.19.0 <4.0.0"
+  flutter: ">=3.7.0"
 
 flutter:
   plugin:
@@ -16,24 +17,39 @@ flutter:
         default_package: image_picker_android
       ios:
         default_package: image_picker_ios
+      linux:
+        default_package: image_picker_linux
+      macos:
+        default_package: image_picker_macos
       web:
         default_package: image_picker_for_web
+      windows:
+        default_package: image_picker_windows
 
 dependencies:
   flutter:
     sdk: flutter
-  image_picker_android: ^0.8.4+11
-  image_picker_for_web: ^2.1.0
+  image_picker_android: ^0.8.7
+  image_picker_for_web: ">=2.2.0 <4.0.0"
   image_picker_ios:
     git:
       url: https://github.com/blinkoo-it/image-picker-ios
-      ref: 0.8.6
-  image_picker_platform_interface: ^2.6.1
+      ref: 0.8.8+2
+  image_picker_linux: ^0.2.1
+  image_picker_macos: ^0.2.1
+  image_picker_platform_interface: ^2.8.0
+  image_picker_windows: ^0.2.1
 
 dev_dependencies:
   build_runner: ^2.1.10
   cross_file: ^0.3.1+1 # Mockito generates a direct include.
   flutter_test:
     sdk: flutter
-  mockito: ^5.0.0
+  mockito: 5.4.1
   plugin_platform_interface: ^2.0.0
+
+topics:
+  - camera
+  - image-picker
+  - files
+  - file-selection
