@@ -1,7 +1,8 @@
-// Mocks generated by Mockito 5.1.0 from annotations
+// Mocks generated by Mockito 5.4.0 from annotations
 // in image_picker/test/image_picker_test.dart.
 // Do not manually edit this file.
 
+// ignore_for_file: no_leading_underscores_for_library_prefixes
 import 'dart:async' as _i4;
 
 import 'package:cross_file/cross_file.dart' as _i5;
@@ -19,11 +20,28 @@ import 'package:mockito/mockito.dart' as _i1;
 // ignore_for_file: prefer_const_constructors
 // ignore_for_file: unnecessary_parenthesis
 // ignore_for_file: camel_case_types
+// ignore_for_file: subtype_of_sealed_class
 
-class _FakeLostData_0 extends _i1.Fake implements _i2.LostData {}
+class _FakeLostData_0 extends _i1.SmartFake implements _i2.LostData {
+  _FakeLostData_0(
+    Object parent,
+    Invocation parentInvocation,
+  ) : super(
+          parent,
+          parentInvocation,
+        );
+}
 
-class _FakeLostDataResponse_1 extends _i1.Fake
-    implements _i2.LostDataResponse {}
+class _FakeLostDataResponse_1 extends _i1.SmartFake
+    implements _i2.LostDataResponse {
+  _FakeLostDataResponse_1(
+    Object parent,
+    Invocation parentInvocation,
+  ) : super(
+          parent,
+          parentInvocation,
+        );
+}
 
 /// A class which mocks [ImagePickerPlatform].
 ///
@@ -35,120 +53,194 @@ class MockImagePickerPlatform extends _i1.Mock
   }
 
   @override
-  _i4.Future<_i2.PickedFile?> pickImage(
-          {_i2.ImageSource? source,
-          double? maxWidth,
-          double? maxHeight,
-          int? imageQuality,
-          _i2.CameraDevice? preferredCameraDevice = _i2.CameraDevice.rear}) =>
+  _i4.Future<_i2.PickedFile?> pickImage({
+    required _i2.ImageSource? source,
+    double? maxWidth,
+    double? maxHeight,
+    int? imageQuality,
+    _i2.CameraDevice? preferredCameraDevice = _i2.CameraDevice.rear,
+  }) =>
       (super.noSuchMethod(
-              Invocation.method(#pickImage, [], {
-                #source: source,
-                #maxWidth: maxWidth,
-                #maxHeight: maxHeight,
-                #imageQuality: imageQuality,
-                #preferredCameraDevice: preferredCameraDevice
-              }),
-              returnValue: Future<_i2.PickedFile?>.value())
-          as _i4.Future<_i2.PickedFile?>);
-
-  @override
-  _i4.Future<List<_i2.PickedFile>?> pickMultiImage(
-          {double? maxWidth, double? maxHeight, int? imageQuality}) =>
-      (super.noSuchMethod(
-              Invocation.method(#pickMultiImage, [], {
-                #maxWidth: maxWidth,
-                #maxHeight: maxHeight,
-                #imageQuality: imageQuality
-              }),
-              returnValue: Future<List<_i2.PickedFile>?>.value())
-          as _i4.Future<List<_i2.PickedFile>?>);
-
-  @override
-  _i4.Future<_i2.PickedFile?> pickVideo(
-          {_i2.ImageSource? source,
-          _i2.CameraDevice? preferredCameraDevice = _i2.CameraDevice.rear,
-          Duration? maxDuration}) =>
-      (super.noSuchMethod(
-              Invocation.method(#pickVideo, [], {
-                #source: source,
-                #preferredCameraDevice: preferredCameraDevice,
-                #maxDuration: maxDuration
-              }),
-              returnValue: Future<_i2.PickedFile?>.value())
-          as _i4.Future<_i2.PickedFile?>);
-
-  @override
-  _i4.Future<_i2.LostData> retrieveLostData() =>
-      (super.noSuchMethod(Invocation.method(#retrieveLostData, []),
-              returnValue: Future<_i2.LostData>.value(_FakeLostData_0()))
-          as _i4.Future<_i2.LostData>);
-
-  @override
-  _i4.Future<_i5.XFile?> getImage(
-          {_i2.ImageSource? source,
-          double? maxWidth,
-          double? maxHeight,
-          int? imageQuality,
-          _i2.CameraDevice? preferredCameraDevice = _i2.CameraDevice.rear}) =>
-      (super.noSuchMethod(
-          Invocation.method(#getImage, [], {
+        Invocation.method(
+          #pickImage,
+          [],
+          {
             #source: source,
             #maxWidth: maxWidth,
             #maxHeight: maxHeight,
             #imageQuality: imageQuality,
-            #preferredCameraDevice: preferredCameraDevice
-          }),
-          returnValue: Future<_i5.XFile?>.value()) as _i4.Future<_i5.XFile?>);
-
+            #preferredCameraDevice: preferredCameraDevice,
+          },
+        ),
+        returnValue: _i4.Future<_i2.PickedFile?>.value(),
+      ) as _i4.Future<_i2.PickedFile?>);
   @override
-  _i4.Future<List<_i5.XFile>?> getMultiImage(
-          {double? maxWidth, double? maxHeight, int? imageQuality}) =>
+  _i4.Future<List<_i2.PickedFile>?> pickMultiImage({
+    double? maxWidth,
+    double? maxHeight,
+    int? imageQuality,
+  }) =>
       (super.noSuchMethod(
-              Invocation.method(#getMultiImage, [], {
-                #maxWidth: maxWidth,
-                #maxHeight: maxHeight,
-                #imageQuality: imageQuality
-              }),
-              returnValue: Future<List<_i5.XFile>?>.value())
-          as _i4.Future<List<_i5.XFile>?>);
-
+        Invocation.method(
+          #pickMultiImage,
+          [],
+          {
+            #maxWidth: maxWidth,
+            #maxHeight: maxHeight,
+            #imageQuality: imageQuality,
+          },
+        ),
+        returnValue: _i4.Future<List<_i2.PickedFile>?>.value(),
+      ) as _i4.Future<List<_i2.PickedFile>?>);
   @override
-  _i4.Future<_i5.XFile?> getVideo(
-          {_i2.ImageSource? source,
-          _i2.CameraDevice? preferredCameraDevice = _i2.CameraDevice.rear,
-          Duration? maxDuration}) =>
+  _i4.Future<_i2.PickedFile?> pickVideo({
+    required _i2.ImageSource? source,
+    _i2.CameraDevice? preferredCameraDevice = _i2.CameraDevice.rear,
+    Duration? maxDuration,
+  }) =>
       (super.noSuchMethod(
-          Invocation.method(#getVideo, [], {
+        Invocation.method(
+          #pickVideo,
+          [],
+          {
             #source: source,
             #preferredCameraDevice: preferredCameraDevice,
-            #maxDuration: maxDuration
-          }),
-          returnValue: Future<_i5.XFile?>.value()) as _i4.Future<_i5.XFile?>);
-
+            #maxDuration: maxDuration,
+          },
+        ),
+        returnValue: _i4.Future<_i2.PickedFile?>.value(),
+      ) as _i4.Future<_i2.PickedFile?>);
   @override
-  _i4.Future<_i2.LostDataResponse> getLostData() =>
-      (super.noSuchMethod(Invocation.method(#getLostData, []),
-              returnValue:
-                  Future<_i2.LostDataResponse>.value(_FakeLostDataResponse_1()))
-          as _i4.Future<_i2.LostDataResponse>);
-
-  @override
-  _i4.Future<_i5.XFile?> getImageFromSource(
-          {_i2.ImageSource? source,
-          _i2.ImagePickerOptions? options = const _i2.ImagePickerOptions()}) =>
-      (super.noSuchMethod(
+  _i4.Future<_i2.LostData> retrieveLostData() => (super.noSuchMethod(
+        Invocation.method(
+          #retrieveLostData,
+          [],
+        ),
+        returnValue: _i4.Future<_i2.LostData>.value(_FakeLostData_0(
+          this,
           Invocation.method(
-              #getImageFromSource, [], {#source: source, #options: options}),
-          returnValue: Future<_i5.XFile?>.value()) as _i4.Future<_i5.XFile?>);
-
+            #retrieveLostData,
+            [],
+          ),
+        )),
+      ) as _i4.Future<_i2.LostData>);
+  @override
+  _i4.Future<_i5.XFile?> getImage({
+    required _i2.ImageSource? source,
+    double? maxWidth,
+    double? maxHeight,
+    int? imageQuality,
+    _i2.CameraDevice? preferredCameraDevice = _i2.CameraDevice.rear,
+  }) =>
+      (super.noSuchMethod(
+        Invocation.method(
+          #getImage,
+          [],
+          {
+            #source: source,
+            #maxWidth: maxWidth,
+            #maxHeight: maxHeight,
+            #imageQuality: imageQuality,
+            #preferredCameraDevice: preferredCameraDevice,
+          },
+        ),
+        returnValue: _i4.Future<_i5.XFile?>.value(),
+      ) as _i4.Future<_i5.XFile?>);
+  @override
+  _i4.Future<List<_i5.XFile>?> getMultiImage({
+    double? maxWidth,
+    double? maxHeight,
+    int? imageQuality,
+  }) =>
+      (super.noSuchMethod(
+        Invocation.method(
+          #getMultiImage,
+          [],
+          {
+            #maxWidth: maxWidth,
+            #maxHeight: maxHeight,
+            #imageQuality: imageQuality,
+          },
+        ),
+        returnValue: _i4.Future<List<_i5.XFile>?>.value(),
+      ) as _i4.Future<List<_i5.XFile>?>);
+  @override
+  _i4.Future<List<_i5.XFile>> getMedia({required _i2.MediaOptions? options}) =>
+      (super.noSuchMethod(
+        Invocation.method(
+          #getMedia,
+          [],
+          {#options: options},
+        ),
+        returnValue: _i4.Future<List<_i5.XFile>>.value(<_i5.XFile>[]),
+      ) as _i4.Future<List<_i5.XFile>>);
+  @override
+  _i4.Future<_i5.XFile?> getVideo({
+    required _i2.ImageSource? source,
+    _i2.CameraDevice? preferredCameraDevice = _i2.CameraDevice.rear,
+    Duration? maxDuration,
+  }) =>
+      (super.noSuchMethod(
+        Invocation.method(
+          #getVideo,
+          [],
+          {
+            #source: source,
+            #preferredCameraDevice: preferredCameraDevice,
+            #maxDuration: maxDuration,
+          },
+        ),
+        returnValue: _i4.Future<_i5.XFile?>.value(),
+      ) as _i4.Future<_i5.XFile?>);
+  @override
+  _i4.Future<_i2.LostDataResponse> getLostData() => (super.noSuchMethod(
+        Invocation.method(
+          #getLostData,
+          [],
+        ),
+        returnValue:
+            _i4.Future<_i2.LostDataResponse>.value(_FakeLostDataResponse_1(
+          this,
+          Invocation.method(
+            #getLostData,
+            [],
+          ),
+        )),
+      ) as _i4.Future<_i2.LostDataResponse>);
+  @override
+  _i4.Future<_i5.XFile?> getImageFromSource({
+    required _i2.ImageSource? source,
+    _i2.ImagePickerOptions? options = const _i2.ImagePickerOptions(),
+  }) =>
+      (super.noSuchMethod(
+        Invocation.method(
+          #getImageFromSource,
+          [],
+          {
+            #source: source,
+            #options: options,
+          },
+        ),
+        returnValue: _i4.Future<_i5.XFile?>.value(),
+      ) as _i4.Future<_i5.XFile?>);
   @override
   _i4.Future<List<_i5.XFile>> getMultiImageWithOptions(
           {_i2.MultiImagePickerOptions? options =
               const _i2.MultiImagePickerOptions()}) =>
       (super.noSuchMethod(
-          Invocation.method(#getMultiImageWithOptions, [], {#options: options}),
-          returnValue:
-              Future<List<_i5.XFile>>.value(<_i5.XFile>[])) as _i4
-          .Future<List<_i5.XFile>>);
+        Invocation.method(
+          #getMultiImageWithOptions,
+          [],
+          {#options: options},
+        ),
+        returnValue: _i4.Future<List<_i5.XFile>>.value(<_i5.XFile>[]),
+      ) as _i4.Future<List<_i5.XFile>>);
+  @override
+  bool supportsImageSource(_i2.ImageSource? source) => (super.noSuchMethod(
+        Invocation.method(
+          #supportsImageSource,
+          [source],
+        ),
+        returnValue: false,
+      ) as bool);
 }
